Migrate specphot checker function to helpers.py
