+++ diffsky/experimental/kernels/tests/test_specphot_kernels_merging.py
@@ -10,44 +10,9 @@ from ....merging import merging_model
 from ....param_utils import diffsky_param_wrapper as dpw
 from ...tests import test_lightcone_generators as tlcg
 from .. import specphot_kernels_merging as sppkm
-from .helpers import check_phot_kern_merging_results
+from .helpers import check_phot_kern_merging_results, check_spec_kern_merging_results
 
 TOL = 1e-8
-
-
-def check_spec_kern_merging_results(spec_kern_results, lc_data):
-    n_gals = lc_data.z_obs.size
-    n_lines = len(lc_data.ssp_data.ssp_emline_wave)
-
-    assert spec_kern_results.linelum_gal.shape == (n_gals, n_lines)
-    assert spec_kern_results.linelum_gal_in_situ.shape == (n_gals, n_lines)
-
-    skip_fields = (
-        "diffstar_info_ms",
-        "diffstar_info_q",
-        "burstiness_info_ms",
-        "burstiness_info_q",
-    )
-    for key in spec_kern_results._fields:
-        if key not in skip_fields:
-            arr = getattr(spec_kern_results, key)
-            try:
-                assert np.all(np.isfinite(arr))
-            except ValueError:
-                raise ValueError(f"{key} is not a flat array")
-
-    msk_cen = lc_data.is_central == 1
-    msk_sat = ~msk_cen
-
-    # Enforce centrals can only get more massive and satellites less massive
-    name = "linelum_gal"
-    x = np.log10(getattr(spec_kern_results, name))
-    y = np.log10(getattr(spec_kern_results, name + "_in_situ"))
-    assert np.all(x[msk_cen] >= y[msk_cen] - TOL)
-    assert np.all(x[msk_sat] <= y[msk_sat] + TOL)
-    # Enforce merging is nontrivial
-    assert np.any(x[msk_cen] > y[msk_cen])
-    assert np.any(x[msk_sat] < y[msk_sat])
 
 
 @pytest.mark.parametrize("mc_merge", [0, 1])
